requirements.txt
amqp==5.1.1
asgiref==3.7.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-celery-results==2.5.1
djangorestframework==3.14.0
google-play-scraper==1.2.4
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.6.2
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
